+++ sc_bp_watcher.py
@@ -59,7 +59,7 @@ try:
 except ImportError:
     winsound = None
 
-__version__ = '3.9.2-rc3'
+__version__ = '3.9.2-rc4'
 
 
 def _mitgeliefert(name):